+++ src/bacnet/basic/object/bacfile.h
@@ -119,16 +119,16 @@ uint32_t bacfile_write(
 
 BACNET_STACK_EXPORT
 void bacfile_write_stream_data_callback_set(
-    size_t (*callback)(const char *, size_t, const uint8_t *, size_t));
+    size_t (*callback)(const char *, int32_t, const uint8_t *, size_t));
 BACNET_STACK_EXPORT
 void bacfile_read_stream_data_callback_set(
-    size_t (*callback)(const char *, size_t, uint8_t *, size_t));
+    size_t (*callback)(const char *, int32_t, uint8_t *, size_t));
 BACNET_STACK_EXPORT
 void bacfile_write_record_data_callback_set(
-    bool (*callback)(const char *, size_t, size_t, const uint8_t *, size_t));
+    bool (*callback)(const char *, int32_t, size_t, const uint8_t *, size_t));
 BACNET_STACK_EXPORT
 void bacfile_read_record_data_callback_set(
-    bool (*callback)(const char *, size_t, size_t, uint8_t *, size_t));
+    bool (*callback)(const char *, int32_t, size_t, uint8_t *, size_t));
 BACNET_STACK_EXPORT
 void bacfile_file_size_callback_set(size_t (*callback)(const char *));
 BACNET_STACK_EXPORT
